+++ salvo/src/lib/BUILD
@@ -1,10 +1,11 @@
 py_library(
   name = "helper_library",
-  data = glob([
-      '*.py',
-  ], allow_empty=False) + 
-  [
-    "//:api",
+  data = [
+    'cmd_exec.py',
+    'docker_image.py',
+    'docker_volume.py',
+    'job_control_loader.py',
+    '//:api',
   ],
   visibility = ["//visibility:public"],
 )
